Added queries to MySQLCRUD
Added queries to the MySQLCRUD class which allow the end user to select rows based on the result of comparisons that are ran on the row's cells.

There are currently three "comparators":
- Equal (MySQL "=")
- NotEqual (MySQL "!=")
- GreaterThan (MySQL ">")
- GreaterThanOrEqualTo (MySQL ">=")
- LesserThan (MySQL "<")
- LesserThanOrEqualTo (MySQL "<=")
- In (MySQL "IN")
- NotIn (MySQL "NOT IN")
- Like (MySQL "LIKE")
- NotLike (MySQL "NOT LIKE")

All of the above comparators, with the exceptions being "Like" and "NotLike" are generic interfaces which take a type parameter.

All these comparators are combined into a single type called "Comparator" this is a generic type which takes a type parameter.